refactor: implement expo router
