+++ src/app/index.tsx
@@ -1,0 +1,17 @@
+import DashboardScreen from 'screens/dashboard';
+import { Stack } from 'expo-router';
+
+export default function Page() {
+    return (
+
+        <>
+            <Stack.Screen
+                options={{
+                    headerTitle: 'Dashboard',
+                }}
+        />
+
+            <DashboardScreen />
+        </>
+    );
+}
