--- src/index.tsx
@@ -1,34 +1,0 @@
-import { NativeBaseProvider, StatusBar } from 'native-base';
-import { Colors } from 'react-native/Libraries/NewAppScreen';
-import { I18nextProvider } from 'react-i18next';
-import { NavigationContainer } from '@react-navigation/native';
-import React from 'react';
-import { useColorScheme } from 'react-native';
-import { customTheme } from '@/helpers/theme';
-import i18n from '@/i18n';
-import App from '@/App';
-
-const Root = (): JSX.Element => {
-    const isDarkMode = useColorScheme() === 'dark';
-
-    const backgroundStyle = {
-        backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
-    };
-
-    return (
-        <NavigationContainer>
-            <I18nextProvider i18n={i18n}>
-                <NativeBaseProvider theme={customTheme}>
-                    <App />
-
-                    <StatusBar
-                        backgroundColor={backgroundStyle.backgroundColor}
-                        barStyle={isDarkMode ? 'light-content' : 'dark-content'}
-                    />
-                </NativeBaseProvider>
-            </I18nextProvider>
-        </NavigationContainer>
-    );
-};
-
-export default Root;
