--- src/router/screens.ts
@@ -1,4 +1,0 @@
-export const screens = {
-    Dashboard: 'Dashboard',
-    About: 'About',
-};
