+++ src/screens/dashboard/index.tsx
@@ -1,20 +1,17 @@
 import { Box, Column, Image, ScrollView, useToast } from 'native-base';
-import type { NavigationProps } from 'router/params';
 import React from 'react';
 import Section from 'components/generic/Section';
 import { useColorScheme } from 'react-native';
-import { useNavigation } from '@react-navigation/native';
+import { router } from 'expo-router';
 import { useTranslation } from 'react-i18next';
 
 const DashboardScreen = () => {
-    const navigation = useNavigation<NavigationProps>();
     const { t } = useTranslation();
     const isDarkMode = useColorScheme() === 'dark';
     const { show } = useToast();
 
     const handleNavigation = () => {
-        // Learn More: https://reactnavigation.org/docs/navigating/
-        navigation.navigate('About');
+        router.push('/about');
     };
 
     const handleNativeBase = () => {
